Feature/Fixed signup bug and Implemented Schedular to delete unverifued OTP autometically

+++ backend/TrueSplit_Backend/TrueSplit_Backend/src/main/java/com/truesplit/TrueSplit/service/OtpService.java
@@ -1,11 +1,14 @@
 package com.truesplit.TrueSplit.service;
 
 import com.truesplit.TrueSplit.Repository.OtpRepository;
+import com.truesplit.TrueSplit.Repository.UserRepository;
 import com.truesplit.TrueSplit.model.Otp;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.MailSender;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -13,25 +16,34 @@ import java.time.Instant;
 import java.util.Random;
 
 @Service
+@Slf4j
 public class OtpService {
     private final OtpRepository otpRepository;
     private final JavaMailSender javaMailSender;
     private final long otpValiditySeconds;
+    private final UserRepository userRepository;
 
     public OtpService(OtpRepository otpRepository, JavaMailSender javaMailSender,
-                      @Value("${otp.validity-seconds:300}") long otpValidityInSeconds) {
+                      @Value("${otp.validity-seconds:60}") long otpValidityInSeconds, UserRepository userRepository) {
         this.otpRepository = otpRepository;
         this.javaMailSender = javaMailSender;
         this.otpValiditySeconds = otpValidityInSeconds;
+        this.userRepository = userRepository;
     }
 
     public String generateOtpAndSend(String email) {
-        String code = String.format("%06d", new Random().nextInt(1_000_000_000));
+
+        if(userRepository.existsByEmail(email)) {
+            throw new IllegalArgumentException("Email already in use");
+        }
+
+        String code = String.format("%06d", new Random().nextInt(1_000_000));
 
         Otp otp = new Otp();
         otp.setEmail(email);
         otp.setCode(code);
-        otp.setExpiresAt(Instant.parse(String.valueOf(Instant.now().plusSeconds(otpValiditySeconds))));
+        //otp.setExpiresAt(Instant.parse(String.valueOf(Instant.now().plusSeconds(otpValiditySeconds))));
+        otp.setExpiresAt(Instant.now().plusSeconds(otpValiditySeconds));
         otpRepository.deleteByEmail(email);
         otpRepository.save(otp);
 
@@ -40,23 +52,35 @@ public class OtpService {
         SimpleMailMessage  message = new SimpleMailMessage();
         message.setTo(email);
         message.setSubject("Your OTP for TrueSplit");
-        message.setText("Your OTP code is: "+ code + "\nIt is valid for " + (otpValiditySeconds/60) + " minutes.");
+        message.setText("Your OTP code is: "+ code + "\nIt is valid for " + (otpValiditySeconds) + " second.");
         javaMailSender.send(message);
         return code;
     }
 
     public boolean verifyOtp(String email, String code) {
         var o = otpRepository.findByEmailAndCode(email, code);
-        if(o.isPresent()) {
-            if(o.get().getExpiresAt().isAfter(Instant.now())) {
-                otpRepository.deleteByEmail(email);
-                return true;
-            }
-            else {
-                otpRepository.deleteByEmail(email);
-                return false;
-            }
+
+        if(o.isEmpty()) return false;
+
+        Otp otp = o.get();
+
+        if (otp.getExpiresAt().isBefore(Instant.now())) {
+            otpRepository.deleteByEmail(email);
+            return false;
         }
-        return false;
+
+        otp.setVerified(true);
+        otpRepository.save(otp); // Save verified
+
+        return true;
+    }
+
+    @Scheduled(fixedRate = 90000) // every 90 seconds
+    public void deleteExpiredOtps() {
+        long deletedCount = otpRepository.deleteByVerifiedFalseAndExpiresAtBefore(Instant.now());
+
+        if (deletedCount > 0) {
+            log.info("[OTP CLEANUP] Deleted {} expired & unverified OTP(s)", deletedCount);
+        }
     }
 }
